sites/funmooc/requirements/base.txt
boto3==1.16.53
django-check-seo==0.4.0
django-configurations==2.2
# Superior versions of django-storages are not compatible with
# ManifestStaticFilesStorage
django-storages==1.11.1
dockerflow==2020.10.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
richie==2.7.1
sentry-sdk==0.19.5
